refactor(primitives): move TextEdit from view/ui/ to primitives/
Step 1 of the Settings + widget framework unification: TextEdit needs to
be shared between the legacy Settings UI (`view/ui/`) and the plugin
widget framework (`widgets/`). Move it to `crates/fresh-editor/src/
primitives/text_edit.rs` so both layers can depend on it without one
reaching into the other's tree.

`crate::view::ui::TextEdit` continues to compile via a re-export from
the new location, so no Settings or controls code had to change beyond
the one direct-path import in `number_input`.
